Make the class diagram generation reproducible
Whilst working on the Reproducible Builds effort [0] I noticed that
sphinxcontrib-mermaid could not be built reproducibly.

This is because the inheritance properties of class diagrams were collated
using a Python set() object, and the contents were not sorted when serialised.

Another alternative would be to replace the set() with a list, but that would
have made this patch more invasive (eg. needing to change .add to .append, etc.
etc.)

I originally filed this in Debian as bug #1012790 [1].

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1012790