Use an <img> for SVG output instead of an <object>
<img>, unlike <object>, has proper alt text support (i.e. that's not
just fallback content) and is easily saved, opened in a new tab, etc.
from the page where it appears.

This does mean the SVG's DOM will no longer be available for inspection
by JavaScript in the containing document (e.g. object.contentDocument),
but I don't think that was intended to be a feature of SVG output for
this Sphinx extension.

Resolves <https://github.com/mgaitan/sphinxcontrib-mermaid/issues/102>
as well.